Refactor stack frame builder into a separate class as requested by reviewer
