+++ torch_xla/csrc/lowering_context.cpp
@@ -16,12 +16,10 @@
 #include "torch_xla/csrc/runtime/debug_macros.h"
 #include "torch_xla/csrc/runtime/sys_util.h"
 #include "torch_xla/csrc/shape_helper.h"
+#include "torch_xla/csrc/stack_frame_index_builder.h"
 #include "torch_xla/csrc/unwrap_data.h"
 
 namespace torch_xla {
-
-// Invalid stack frame id - used for stack frame population
-static int kInvalidIndex = 0;
 
 namespace {
 
@@ -79,24 +77,9 @@ class HloMetadataSetter {
     }
     metadata.set_op_name(absl::StrCat(op_name_prefix, op_type));
 
-    if (!nmeta.frame_info.empty()) {
-      auto frame_it = nmeta.frame_info.rbegin();
-      int parent_frame_id = kInvalidIndex;
-      int depth = 0;
-      for (; frame_it != nmeta.frame_info.rend() && depth < max_stack_depth;
-           ++frame_it) {
-        parent_frame_id =
-            loctx->AddStackFrameLocation(*frame_it, parent_frame_id);
-        ++depth;
-      }
-
-      // Point to first entry / deepest call / top frame in call stack
-      --frame_it;
-
-      metadata.set_source_file(frame_it->file);
-      metadata.set_source_line(frame_it->line);
-      metadata.set_stack_frame_id(parent_frame_id);
-    }
+    // Sets file, line and stack_frame_id in metadata
+    loctx->stack_frame_index_builder()->AddStackFrameLocations(
+        nmeta.frame_info, max_stack_depth, metadata);
 
     loctx->builder()->SetOpMetadata(std::move(metadata));
   }
@@ -106,25 +89,19 @@ class HloMetadataSetter {
 
 }  // namespace
 
-int FindId(std::string_view key, std::map<std::string_view, int>& index) {
-  auto entry_iterator = index.find(key);
-  if (entry_iterator == index.end()) {
-    return 0;
-  } else {
-    return entry_iterator->second;
-  }
-}
-
 LoweringContext::LoweringContext(const std::string& name,
                                  torch::lazy::BackendDevice device)
-    : torch::lazy::LoweringContext(name, device), builder_(name) {}
+    : torch::lazy::LoweringContext(name, device),
+      builder_(name),
+      stack_frame_index_builder_(std::make_shared<StackFrameIndexBuilder>()) {}
 
 LoweringContext::LoweringContext(
     const std::string& name, torch::lazy::BackendDevice device,
     c10::ArrayRef<const torch::lazy::Node*> post_order,
     torch::lazy::Util::EmissionMap emit_status)
     : torch::lazy::LoweringContext(name, device, {}, emit_status),
-      builder_(name) {
+      builder_(name),
+      stack_frame_index_builder_(std::make_shared<StackFrameIndexBuilder>()) {
   for (auto node : post_order) {
     LowerNode(node);
   }
@@ -187,7 +164,8 @@ xla::StatusOr<xla::XlaComputation> LoweringContext::BuildXla() {
   }
 
   if (xla.ok()) {
-    (*xla->mutable_proto()->mutable_stack_frame_index()) = indexes_;
+    (*xla->mutable_proto()->mutable_stack_frame_index()) =
+        stack_frame_index_builder()->stack_frame_index();
   }
 
   return xla;
@@ -198,7 +176,8 @@ xla::StatusOr<xla::XlaComputation> LoweringContext::BuildXla(xla::XlaOp root) {
   auto xla = builder()->Build(root);
 
   if (xla.ok()) {
-    (*xla->mutable_proto()->mutable_stack_frame_index()) = indexes_;
+    (*xla->mutable_proto()->mutable_stack_frame_index()) =
+        stack_frame_index_builder()->stack_frame_index();
   }
 
   return xla;
@@ -310,63 +289,6 @@ void LoweringContext::AddParameter(const torch::lazy::Output& output,
                                    const std::string& name) {
   XLA_ERROR() << "not implemented";
   return;
-}
-
-// TODO: Remove duplicate code copied form openxla/xla MHLO -> HLO
-int64_t LoweringContext::AddStackFrameLocation(
-    const torch::lazy::SourceLocation& frame, int64_t parent_frame_id) {
-  int line = frame.line;
-  int column = 0;  // Not provided in torch lazy source location - set to zero
-  std::string filename = frame.file;
-  std::string function_name = frame.function;
-
-  int filename_id = FindId(filename, file_name_to_id_);
-  if (filename_id == 0) {
-    indexes_.add_file_names(std::move(filename));
-    filename_id = indexes_.file_names_size();
-    file_name_to_id_[indexes_.file_names(filename_id - 1)] = filename_id;
-  }
-
-  int function_name_id = FindId(function_name, function_name_to_id_);
-  if (function_name_id == 0) {
-    indexes_.add_function_names(std::move(function_name));
-    function_name_id = indexes_.function_names_size();
-    function_name_to_id_[indexes_.function_names(function_name_id - 1)] =
-        function_name_id;
-  }
-
-  auto location_tuple =
-      std::make_tuple(filename_id, function_name_id, line, column);
-  auto file_location_iterator = file_location_to_id_.find(location_tuple);
-  int file_location_id = 0;
-  if (file_location_iterator == file_location_to_id_.end()) {
-    auto file_location = indexes_.add_file_locations();
-    file_location->set_file_name_id(filename_id);
-    file_location->set_function_name_id(function_name_id);
-    file_location->set_line(line);
-    file_location->set_column(column);
-
-    file_location_id = indexes_.file_locations_size();
-    file_location_to_id_[location_tuple] = file_location_id;
-  } else {
-    file_location_id = file_location_iterator->second;
-  }
-
-  auto frame_tuple = std::make_tuple(file_location_id, parent_frame_id);
-  auto stack_frame_iterator = frame_to_id_.find(frame_tuple);
-  int stack_frame_id = 0;
-  if (stack_frame_iterator == frame_to_id_.end()) {
-    auto frame = indexes_.add_stack_frames();
-    frame->set_file_location_id(file_location_id);
-    frame->set_parent_frame_id(parent_frame_id);
-
-    stack_frame_id = indexes_.stack_frames_size();
-    frame_to_id_[frame_tuple] = stack_frame_id;
-  } else {
-    stack_frame_id = stack_frame_iterator->second;
-  }
-
-  return stack_frame_id;
 }
 
 torch::lazy::ComputationPtr LoweringContext::Build() {
